requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.23.0
python-jose>=3.3.0
slowapi>=0.1.5
redis>=4.5.0
prometheus-client>=0.17.0
